chore: refactored QiwiWalletComponent and MBWayComponent to use a base class PhoneBasedPaymentComponent to contain the common code
